pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xldvp_seg"
version = "2.0.0"
description = "Spatial cell segmentation and DVP pipeline for CZI microscopy (8 detection strategies, multi-GPU, LMD export)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.12"
authors = [
    {name = "Cell Segmentation Team"}
]
keywords = ["segmentation", "microscopy", "cells", "deep-learning", "SAM2", "cellpose"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
]

# Core dependencies (PyTorch and SAM2 installed separately - see install.sh)
dependencies = [
    "numpy>=1.24",
    "pandas>=1.5",
    "opencv-python>=4.8",
    "scipy>=1.10",
    "scikit-learn>=1.0",
    "scikit-image>=0.21",
    "h5py>=3.8",
    "hdf5plugin>=4.0",
    "tqdm>=4.65",
    "pillow>=9.5",
    "aicspylibczi>=3.1",
    "matplotlib>=3.7",
    "psutil>=5.8",
    "joblib>=1.2",
    "cellpose>=3.0",
    # SpatialData / scverse ecosystem
    "spatialdata>=0.1",
    "anndata>=0.10",
    "scanpy>=1.9",
    "squidpy>=1.3",
    "geopandas>=0.14",
    "shapely>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black==25.12.0",
    "ruff==0.15.7",
]
brightfield = [
    "timm>=0.9",
    "transformers>=4.30",
    "huggingface-hub>=0.20",
]
instanseg = [
    "instanseg-torch>=0.1",
]

[project.scripts]
xlseg = "xldvp_seg.cli.main:cli"

[project.urls]
"Homepage" = "https://github.com/peptiderodriguez/xldvp_seg"
"Repository" = "https://github.com/peptiderodriguez/xldvp_seg"
"Bug Tracker" = "https://github.com/peptiderodriguez/xldvp_seg/issues"
"Documentation" = "https://github.com/peptiderodriguez/xldvp_seg/blob/main/docs/GETTING_STARTED.md"

[tool.setuptools]
py-modules = [
    "run_segmentation",
    "run_lmd_export",
    "serve_html",
    "train_classifier",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["xldvp_seg*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json"]

[tool.black]
line-length = 100
target-version = ['py310']
extend-exclude = "archive"

[tool.ruff]
line-length = 100
extend-exclude = ["archive/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = [
    "E501",   # line too long (handled by black)
    "E402",   # module-level import not at top (needed for os.environ before imports)
    "E731",   # lambda assignment (used in closures)
    "E741",   # ambiguous variable name (l, I, O — common in scientific code)
    "N803",   # argument name should be lowercase (X, SAM2_model in scientific code)
    "N806",   # variable name should be lowercase (X, NMJ_FEATURES in scientific code)
    "N816",   # mixed-case variable in global scope
    "B007",   # unused loop variable (common in enumerate patterns)
    "B905",   # zip without strict= (Python 3.10 compat)
    "B904",   # raise from within except (verbose for little gain)
    "B006",   # mutable default argument in function (intentional in some cases)
    "B027",   # empty method body (abstract methods)
    "C408",   # unnecessary dict() call (cosmetic)
    "C401",   # unnecessary generator within set() (cosmetic)
    "C405",   # unnecessary literal within set() (cosmetic)
    "C414",   # unnecessary list() within sorted() (cosmetic)
    "C416",   # unnecessary comprehension (cosmetic)
    "F841",   # local variable assigned but never used (common in unpacking)
    "N802",   # function name should be lowercase (scientific code)
]

[tool.ruff.lint.per-file-ignores]
"xldvp_seg/__init__.py" = ["F401"]  # re-exports
"xldvp_seg/processing/deduplication.py" = ["F401"]  # hdf5plugin side-effect import
"xldvp_seg/core/slide_analysis.py" = ["F821"]  # lazy import type hint
"xldvp_seg/detection/strategies/tissue_pattern.py" = ["F821"]  # forward reference
"tests/test_deduplication.py" = ["F401"]  # hdf5plugin side-effect import
"tests/test_feature_extraction.py" = ["F401"]  # torch availability check

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
markers = [
    "integration: needs GPU + real CZI data (run on cluster, not CI)",
]
